archive/cleanup_2025_07_20/legacy_backup/backup_before_cleanup/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
pyarrow>=10.0.0
scipy>=1.9.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
lru-dict>=1.1.8
pathlib2>=2.3.0
pydantic>=2.0.0

# Production dependencies
aiohttp>=3.8.0
aioredis>=2.0.0
aiokafka>=0.8.0
redis>=4.0.0
kafka-python>=2.0.0
websockets>=10.0
